refactor: 테스트 적용한 native-compass 플러그인을 커스텀 후크로 분리
